+++ graphite.c
@@ -61,7 +61,7 @@ int graphite_write(struct s_graphite_entry *entry) {
 	(void)snprintf(buf, sizeof(buf), "%s.temperature %.02f %s\n", gc->namespace, entry->temperature, entry->timestamp);
 	write(s, &buf, strlen(buf));
 
-	(void)snprintf(buf, sizeof(buf), "%s.pressure %d %s\n", gc->namespace, (unsigned int)entry->pressure, entry->timestamp);
+	(void)snprintf(buf, sizeof(buf), "%s.pressure %.02f %s\n", gc->namespace, entry->pressure, entry->timestamp);
 	write(s, &buf, strlen(buf));
 
 	(void)snprintf(buf, sizeof(buf), "%s.humidity %.02f %s\n", gc->namespace, entry->humidity, entry->timestamp);
