+++ packet.h
@@ -22,7 +22,7 @@ struct s_packet {
 	float temperature;
 	float humidity;
 	float rainfall;
-	long pressure;
+	float pressure;
 	float wind_speed;
 	long wind_direction;
 	float light;
